JsonResponse
JsonResponse is a JSON variant of the PSR ResponseInterface.

Unlike JsonRequest we don't need to inherit we can just implement the
same interface, mostly with pass through delegation.

The main addition is passing in the JSON data and access to the getData
method to get it back out again.

I was going to implement this as a public property but it turned out to
be desirable to be able to mock it.