JsonClient
JsonClient provides the final thin layer over the top of Guzzle's
Client.

We take a JsonRequest and return a JsonResponse after decoding the body
we get from Guzzle's own send.

Empty body is encoded as null data.  It's possible we could've encoded
this as an empty array but I can't tell if the distinction will be
useful so I'm living with it being nullable.